+++ gradle-guide-error-prone/src/main/java/com/palantir/gradle/guide/errorprone/ProviderGet.java
@@ -25,6 +25,7 @@ import com.google.errorprone.fixes.SuggestedFix;
 import com.google.errorprone.matchers.Description;
 import com.google.errorprone.matchers.Matcher;
 import com.google.errorprone.matchers.Matchers;
+import com.palantir.gradle.guide.errorprone.utils.TreeUtils;
 import com.sun.source.tree.ExpressionTree;
 import com.sun.source.tree.LambdaExpressionTree;
 import com.sun.source.tree.MemberSelectTree;
